mini-projet/auth-service/requirements.txt
Flask==3.0.3
PyJWT==2.9.0
Werkzeug==3.0.3
